+++ src/pages/ItemList/index.tsx
@@ -1,29 +1,23 @@
-import React, { useRef } from "react";
-import { useNavigate } from "react-router-dom";
+import { useRef } from "react";
 import GridWrapper from "../../components/GridWrapper";
-import Item from "../../components/Item";
 import ItemSkeleton from "../../components/ItemSkeleton";
 import { v4 as uuidv4 } from "uuid";
 import throttle from "../../utils/throttle";
 import useInfinityScroll from "../../hooks/useInfinityScroll";
 import { LOAD_ITEM_AMOUNT } from "../../constants";
 import useProductList from "../../hooks/useProductList";
-import { product } from "../../types/product";
-import useCart from "../../hooks/useCart";
+import ProductList from "../../components/ProductList";
 
 const DELAY_TIME = 500;
 
 const ItemList = () => {
   const sectionRef = useRef(null);
-  const { createNewCart } = useCart();
   const {
     isLoading,
     data: products,
     isEnd,
     getProductsByPage,
   } = useProductList();
-
-  const navigate = useNavigate();
 
   const delayGetProduct = throttle(DELAY_TIME, () => getProductsByPage());
 
@@ -32,27 +26,6 @@ const ItemList = () => {
     cb: delayGetProduct,
     endPoint: isEnd,
   });
-
-  const handleItemClick = (id: number) => {
-    navigate(`/product/${id}`);
-  };
-
-  const ProductList: Function = (products: product[]): React.ReactElement[] =>
-    products.map((product: product) => {
-      const { id } = product;
-      const itemCardElemProps = {
-        key: product.id,
-        onClick: () => {
-          handleItemClick(product.id);
-        },
-        onClickShoppingCart: () => {
-          createNewCart(id, product.price);
-        },
-        ...product,
-      };
-
-      return <Item {...itemCardElemProps} />;
-    });
 
   return (
     <section>
